-started adding exporting -changed most (like 99%) of floats to doubles in block based coding -added click event to blocks -added option for custom block blueprints (BlockFactory) -changed workflow a little for creating blocks from blockoptions, there are not more intermediate steps -added block "registry" -fixed FillingComponent not respecting minHeight -started adding string input block -easter egg from annon -added (i think wrong) a min drag dist (less than that is a click)

+++ src/main/java/com/kyfexuwu/m3we/editor/BlockDrawHelper.java
@@ -17,8 +17,8 @@ public class BlockDrawHelper {
             this.matrix=matrix;
             this.color=color;
         }
-        public void vertex(float x, float y){
-            this.color.setColor(this.builder.vertex(this.matrix, x, y, 0.0f)).next();
+        public void vertex(double x, double y){
+            this.color.setColor(this.builder.vertex(this.matrix, (float)x, (float)y, 0.0f)).next();
         }
     }
     public static void vertexes(MatrixStack matrices, Color color, Consumer<Context>... funcs){
